+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 All notable changes to this project will be documented in this file.
 We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 
+## Unreleased
+- Added a script to tag settlement schools using an S3 csv of school IDs
+- Added manage command to run update_national_stats_file
+
 ## 2.2.2
 - Switch to djangorestframework==2.4.8 to match servers